refactor: 포맷용 모듈 분리, 주석 추가
