+++ common/database.js
@@ -50,6 +50,9 @@ function updateItem(fileName, id, updatedItem) {
   writeJson(fileName, updatedData);
 }
 
+/**
+ * 데이터 삭제 함수 추가 (ID 기반)
+ */
 function deleteItem(fileName, id) {
   const data = readJson(fileName);
   const updatedData = data.filter(item => item.id !== id);
