+++ data/schedules.json
@@ -41,18 +41,10 @@
   },
   {
     "id": 1745689009947,
-    "title": "",
+    "title": "safsafddd",
     "userId": "U08MYM03PFG",
     "createdAt": "2025-04-26T17:36:49.947Z",
-    "dueDate": "2020/02/02",
-    "isCompleted": false
-  },
-  {
-    "id": 1745779467237,
-    "title": "sd",
-    "userId": "U08MYM03PFG",
-    "createdAt": "2025-04-27T18:44:27.237Z",
-    "dueDate": "2222/02/02",
+    "dueDate": "2021/02/03",
     "isCompleted": false
   },
   {
